renamed paths for planDetail components and styling
